@@ -7,35 +7,35 @@
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.2.2",
     "@use-hooks/axios": "^1.3.1",
-    "antd": "4.8.5",
+    "antd": "5.27.4",
     "axios": "^0.21.1",
-    "chart.js": "^2.9.4",
-    "chartjs-plugin-datalabels": "^0.7.0",
-    "chartjs-plugin-piechart-outlabels": "^0.1.4",
+    "chart.js": "^4.5.0",
     "color-hash": "^2.0.1",
     "dayjs": "^1.9.7",
     "framer-motion": "^12.23.12",
+    "highlight-words-core": "^1.2.3",
     "html2canvas": "^1.0.0-rc.7",
     "jspdf": "^2.3.0",
     "jwt-decode": "^3.1.2",
-    "node-sass": "^4.14.1",
-    "react": "^17.0.1",
-    "react-chartjs-2": "^2.11.1",
+    "memoize-one": "^6.0.0",
+    "moment": "^2.30.1",
+    "path-browserify": "^1.0.1",
+    "react": "^18.3.1",
+    "react-chartjs-2": "^5.3.0",
     "react-csv": "^2.0.3",
-    "react-dom": "^17.0.1",
-    "react-highlight-words": "^0.16.0",
-    "react-hook-form": "^6.13.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.64.0",
     "react-images-uploading": "^3.1.3",
     "react-onclickoutside": "^6.9.0",
     "react-redux": "^7.2.5",
-    "react-responsive": "^8.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "recharts": "2.8.0",
+    "react-responsive": "^10.0.1",
+    "react-router-dom": "^7.9.3",
+    "react-scripts": "^5.0.1",
+    "recharts": "^3.2.1",
     "redux": "^4.1.1",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0",
-    "sass": "^1.34.0",
+    "sass": "^1.93.0",
     "uuid": "^8.3.2",
     "web-vitals": "^0.2.4",
     "xlsx": "^0.18.5"
@@ -76,7 +76,6 @@
   },
   "devDependencies": {
     "@iconify/react": "^4.1.0",
-    "json-server": "^0.16.3",
-    "redux-devtools": "^3.7.0"
+    "json-server": "^0.16.3"
   }
 }
