@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
 // import Home from '../authPages/Home';
 import About from '../authPages/About';
@@ -17,24 +17,25 @@ import AuthHeader from '../components/AuthHeader';
 import Footer from '../components/Footer';
 
 function AuthPages() {
+
   return (
     <div>
       <AuthHeader />
 
       <main className='auth-container'>
-        <ScrollToTop>
-          <Switch>
-            <Route exact path='/' component={Login} />
-            <Route path='/about' component={About} />
-            <Route path='/contact' component={Contact} />
-            <Route path='/features' component={Features} />
-            <Route path='/log-in' component={Login} />
-            <Route path='/sign-up' component={SignUp} />
-            <Route path='/change-password' component={ChangePassword} />
-            <Route path='/reset-password' component={ResetPassword} />
-            <Route component={Error} />
-          </Switch>
-        </ScrollToTop>
+        {/* <ScrollToTop> */}
+          <Routes> 
+            <Route path="/" element={<Login to="/" replace />} />           
+            <Route path='/about' element={<About to="/about" /> } />
+            <Route path='/contact' element={<Contact to="/contact" /> } />
+            <Route path='/features' element={<Features to="/features" />} />
+            <Route path='/log-in' element={<Login to="/log-in" />}  />
+            <Route path='/sign-up' element={<SignUp to="/sign-up" /> } />
+            <Route path='/change-password' element={<ChangePassword to="/change-password" /> } />
+            <Route path='/reset-password' element={<ResetPassword to="/reset-password" /> } />
+            <Route path="*" element={<Error />} />
+          </Routes>
+        {/* </ScrollToTop> */}
       </main>
 
       {/* <Footer /> */}
@@ -43,3 +44,4 @@ function AuthPages() {
 }
 
 export default AuthPages;
+
