@@ -3,7 +3,6 @@ import React, { useState, useContext } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import { Spin, notification } from 'antd';
 import { useForm, Controller } from 'react-hook-form';
-import jwt from 'jwt-decode';
 
 import CompleteDataContext from '../Context';
 import loginHttpServices from '../services/login';
@@ -31,6 +30,7 @@ function Login() {
   const { register, handleSubmit, control } = useForm();
 
   const onSubmit = async ({ username, password }, values) => {
+    removeErrorMessage('')
     try {
       setIsAuthenticating(true);
       localStorage.clear();
@@ -41,7 +41,7 @@ function Login() {
       });
 
       window.localStorage.setItem('loggedWyreUser', JSON.stringify(user.data.token));
-      
+
       window.location.href = from;
       // dataHttpServices.setUserId(user.data.id);
       // dataHttpServices.setToken(user.data.token);
@@ -55,6 +55,14 @@ function Login() {
       })
     }
   };
+  const onSubmitt = (data, e) => console.log('success', data, e);
+  // const onError = (errors, e) => console.log('error', errors, e);
+
+  const onError = (errors, e) => {
+    console.log('this is the errors', errors)
+    const errorMsg = Object.values(errors)[0]?.message || '';
+    setErrorMessage(errorMsg, errors)
+  }
 
   const removeErrorMessage = (e) => {
     setErrorMessage(undefined);
@@ -63,60 +71,99 @@ function Login() {
   return (
     <div className='auth-page-container'>
       <Spin spinning={isAuthenticating} >
-      <form
-        className='signup-login-contact-form'
-        action='#'
-        onSubmit={handleSubmit(onSubmit)}
-      >
-        <h1 className='signup-login-heading first-heading--auth'>
-          Welcome Back
-        </h1>
+        <form
+          className='signup-login-contact-form'
+          action='#'
+          onSubmit={handleSubmit(onSubmit, onError)}
+        >
+          <h1 className='signup-login-heading first-heading--auth'>
+            Welcome Back
+          </h1>
 
-        <p className='outlined-input-container'>
-          <HiddenInputLabel htmlFor='username' labelText='Username' />
-          <OutlinedInput
-            className='signup-login-contact-input '
-            type='text'
-            name='username'
-            id='username'
-            placeholder='Username'
-            autoComplete='username'
-            required={true}
-            autoFocus={true}
-            register={register}
-            onChange={removeErrorMessage}
-          />
-        </p>
+          <p className='outlined-input-container'>
+            <HiddenInputLabel htmlFor='username' labelText='Username' />
+            <OutlinedInput
+              // defaultValue=''
+              className='signup-login-contact-input '
+              type='text'
+              placeholder='Username'
+              autoComplete='username'
+              register={register('username').ref}
+              onChange={removeErrorMessage}
+              {...register("username", { minLength: 3, required: true })}
+            />
+          </p>
 
-        <p className='outlined-input-container'>
-          <HiddenInputLabel htmlFor='password' labelText='Password' />
-          <Controller as={<Input.Password/>} control={control} defaultValue=''
-                className='signup-login-contact-input outlined-input'
-                type='password'
-                name='password'
-                id='password'
-                placeholder='Password'
-                autoComplete='new-password'
-                rules={{
-                  required : true
-                }}
-                autoFocus={false}
-                onChange={removeErrorMessage}
-                iconRender={visible => (visible ? <EyeTwoTone /> : <EyeInvisibleOutlined />)}/>
-        </p>
+          <p className='outlined-input-container'>
+            <HiddenInputLabel htmlFor='password' labelText='Password' />
+            {/* <Controller
+              // as={<Input.Password />}
+              control={control}
+              // defaultValue=''
+              // rules={{
+              //   required: true
+              // }}
+              // type='password'
+              name='password'
+              // id='password'
+              // {...register('password')}
+              render={({ field }) => (
+                <Input.Password
+                  className='signup-login-contact-input outlined-input'
+                // type='password'
+                // name='password'
+                // id='password'
+                // placeholder='Password'
+                // autoComplete='new-password'
+                // rules={{
+                //   required: true
+                // }}
+                // value={value}
+                // autoFocus={false}
+                // onChange={(e) => onChange(e?.toString())}
+                // onChange={removeErrorMessage}
+                // iconRender={visible => (visible ? <EyeTwoTone /> : <EyeInvisibleOutlined />)}
 
-        <p className='signup-login-contact-error-message'>{errorMessage}</p>
+                />
+              )}
 
-        <div className='forgot-password-wrapper'>
-          <Link className='forgot-password' to='/reset-password'>
-            Forgot Password?
-          </Link>
-        </div>
+            /> */}
 
-        <button className='signup-login-contact-button'>Log in</button>
-      </form>
-        
-      <SocialCluster />
+            <Controller
+              control={control}
+              name="password"
+              rules={{
+                required: true,
+                minLength: {
+                  value: 5,
+                  message: "password must be minimum of 5 characters",
+                },
+              }}
+              onChange={removeErrorMessage}
+              render={({ field: { onChange, onBlur, value, ref } }) => (
+                <Input.Password
+                  className='signup-login-contact-input outlined-input'
+                  onBlur={onBlur}
+                  selected={value}
+                  onChange={onChange}
+                  // onChange={removeErrorMessage}
+                />
+              )}
+            />
+          </p>
+
+          <p className='signup-login-contact-error-message'>{errorMessage}</p>
+
+          <div className='forgot-password-wrapper'>
+            <Link className='forgot-password' to='/reset-password'>
+              Forgot Password?
+            </Link>
+          </div>
+
+          <button className='signup-login-contact-button'>Log in</button>
+        </form>
+
+        <SocialCluster />
       </Spin>
     </div>
   );
