@@ -31,7 +31,7 @@ const ReportPowerDemandGroupedBarChart = ({ data }) => {
     },
     maintainAspectRatio: false,
     scales: {
-      yAxes: [
+      yAxes: 
         {
           stacked: false,
           display: true,
@@ -57,8 +57,7 @@ const ReportPowerDemandGroupedBarChart = ({ data }) => {
             fontColor: 'black',
           },
         },
-      ],
-      xAxes: [
+      xAxes: 
         {
           stacked: false,
           ticks: {
@@ -80,7 +79,6 @@ const ReportPowerDemandGroupedBarChart = ({ data }) => {
             fontColor: 'black',
           },
         },
-      ],
     },
   };
 
