@@ -30,7 +30,7 @@ const RunningTime = ({ runningTimeData, dataMessage }) => {
 
     maintainAspectRatio: false,
     scales: {
-      yAxes: [
+      yAxes: 
         {
           gridLines: {
             display: false,
@@ -51,8 +51,7 @@ const RunningTime = ({ runningTimeData, dataMessage }) => {
             fontSize: isMediumScreen ? 14 : 18,
           },
         },
-      ],
-      xAxes: [
+      xAxes: 
         {
           gridLines: {
             display: false,
@@ -71,7 +70,6 @@ const RunningTime = ({ runningTimeData, dataMessage }) => {
             fontSize: isMediumScreen ? 14 : 18,
           },
         },
-      ],
     },
   };
 
