@@ -1,6 +1,6 @@
 import React from 'react';
 import { Table, Input, Button, Space } from 'antd';
-import Highlighter from 'react-highlight-words';
+import Highlighter from '../Highlighter';
 import { SearchOutlined } from '@ant-design/icons';
 
 import { numberFormatter } from "../../helpers/numberFormatter"
