@@ -1,6 +1,6 @@
 import React from 'react';
 import { Table, Input, Button, Space, Typography } from 'antd';
-import Highlighter from 'react-highlight-words';
+import Highlighter from '../Highlighter';
 import { SearchOutlined } from '@ant-design/icons';
 
 const { Text } = Typography;
