@@ -1,9 +1,9 @@
 import React from 'react';
 import { Table, Input, Button, Space } from 'antd';
-import Highlighter from 'react-highlight-words';
 import { SearchOutlined } from '@ant-design/icons';
 
 import CaretDown from '../../../icons/CaretDown';
+import Highlighter from '../../Highlighter';
 
 class UserBranchesTable extends React.Component {
   state = {
