@@ -53,6 +53,7 @@ function Dashboard({
   fetchPowerFactor: fetchAllPowerFactor,
   fetchPAPR,
   dashboard,
+  powerFactor,
 }) {
   let {
     checkedItems,
@@ -63,8 +64,8 @@ function Dashboard({
     userDateRange,
     uiSettings,
   } = useContext(CompleteDataContext);
-  
-  
+
+
   const dashBoardInfo = useSelector((state) => state.dashboard);
   const branchId =
     sideDetails?.sideBarData?.branches &&
@@ -89,35 +90,48 @@ function Dashboard({
   }, [match, setCurrentUrl]);
 
   useEffect(() => {
-    if (Object.keys(sideDetails.sideBarData).length > 0) {
-      let allDevices = [];
-      sideDetails.sideBarData.branches.forEach((branch) => {
-        branch.devices.forEach((device) => {
-          allDevices.push(device.device_id);
+    if (!(powerFactor.allPowerFactor.length > 0)) {
+      if (Object.keys(sideDetails.sideBarData).length > 0) {
+        let allDevices = [];
+        sideDetails.sideBarData.branches.forEach((branch) => {
+          branch.devices.forEach((device) => {
+            allDevices.push(device.device_id);
+          });
         });
-      });
-      const start_date = moment().startOf("month").format("YYYY-MM-DD");
-      const end_date = moment().startOf("month").format("YYYY-MM-DD");
-      fetchAllPowerFactor(allDevices, { start_date, end_date });
+        const start_date = moment().startOf("month").format("YYYY-MM-DD");
+        const end_date = moment().startOf("month").format("YYYY-MM-DD");
+        fetchAllPowerFactor(allDevices, { start_date, end_date });
+      }
     }
-      fetchBlendedost(userDateRange);
-  }, [sideDetails.sideBarData, userDateRange]);
+
+  }, [sideDetails.sideBarData]);
+
 
   useEffect(() => {
-    if (!pageLoaded && isEmpty(dashBoardInfo.dashBoardData || {})) {
-      fetchDashBoardDataCard_1(userDateRange);
-      fetchDashBoardDataCard_2(userDateRange);
-      fetchDashBoardDataCard_3(userDateRange);
-      fetchPAPR(userDateRange)
+
+    fetchBlendedost(userDateRange);
+
+  }, [userDateRange]);
+
+  useEffect(() => {
+
+    if (!dashboard.fetchDashBoardCard_1_Loading && !dashboard.fetchDashBoardCard_2_Loading && !dashboard.fetchDashBoardCard_3_Loading) {
+      if (!pageLoaded && isEmpty(dashBoardInfo.dashBoardData || {})) {
+        fetchDashBoardDataCard_1(userDateRange);
+        fetchDashBoardDataCard_2(userDateRange);
+        fetchDashBoardDataCard_3(userDateRange);
+        fetchPAPR(userDateRange)
+      }
+
+      if (!isEmpty(dashBoardInfo.dashBoardData) > 0 && pageLoaded) {
+        fetchDashBoardDataCard_1(userDateRange);
+        fetchDashBoardDataCard_2(userDateRange);
+        fetchDashBoardDataCard_3(userDateRange);
+        fetchPAPR(userDateRange)
+      }
+      setPageLoaded(true);
     }
 
-    if (!isEmpty(dashBoardInfo.dashBoardData) > 0 && pageLoaded) {
-      fetchDashBoardDataCard_1(userDateRange);
-      fetchDashBoardDataCard_2(userDateRange);
-      fetchDashBoardDataCard_3(userDateRange);
-      fetchPAPR(userDateRange)
-    }
-    setPageLoaded(true);
   }, [userDateRange]);
 
   useEffect(() => {
@@ -218,7 +232,7 @@ function Dashboard({
                 <Tooltip
                   placement="top"
                   style={{ textAlign: "right" }}
-                  overlayStyle={{ whiteSpace: "pre-line" }}
+                  popupStyle={{ whiteSpace: "pre-line" }}
                   title={DASHBOARD_TOOLTIP_MESSAGES.MAX_MIN_AVERAGE}
                 >
                   <p>
@@ -286,12 +300,14 @@ function Dashboard({
                 })}
           </div>
         </Spin>
+        <Spin spinning={!totalDailyConsumptionBranchData} >
+          <DailyConsumption
+            totalDailyConsumptionBranchData={totalDailyConsumptionBranchData}
+            uiSettings={uiSettings}
+            sideDetails={sideDetails}
+          />
+        </Spin>
 
-        <DailyConsumption
-          totalDailyConsumptionBranchData={totalDailyConsumptionBranchData}
-          uiSettings={uiSettings}
-          sideDetails={sideDetails}
-        />
 
         <div className="dashboard-row-3">
           <PowerUsageCard
@@ -319,13 +335,13 @@ function Dashboard({
                   pDemand={pDemand}
                   runningPercentageData={
                     dashboard.dashBoardCard_1_Data.branches.length > 1 &&
-                    (!checkedItems || Object.keys(checkedItems).length === 0)
+                      (!checkedItems || Object.keys(checkedItems).length === 0)
                       ? generateMultipleBranchLoadOverviewChartData(
-                          dashboard.dashBoardCard_1_Data.branches
-                        )
+                        dashboard.dashBoardCard_1_Data.branches
+                      )
                       : generateLoadOverviewChartData(
-                          dashboard.dashBoardCard_1_Data.branches[0].devices
-                        )
+                        dashboard.dashBoardCard_1_Data.branches[0].devices
+                      )
                   }
                   dataTitle="Operating Time"
                 />
